==> Code/03_Census_Climate_Calculations.qmd <==
---
title: "03. Interval Climate Calculations"
format:
  html:
    theme: spacelab
    toc: true
    number-sections: true
---


 Mortality - disturbance project Balkans
 Calculate census interval mean climate for each plot
 calculate climate anomalies (interval - antecedent)
 
# Set up
```{r}
rm(list=ls())

#  set user computer
computer <- "~"

#  directores
if(computer == "~"){
  Dir.Base <- "~/Desktop/Prague/mort-likelihood/individ-tree-mort"
}else{
  Dir.Base <- "insert path"
}

Dir.Data <- file.path(Dir.Base, "Data")
Dir.Code <- file.path(Dir.Base, "Code")
Dir.Clim <- file.path(Dir.Data,"Climate")
dirs <- sapply(c(Dir.Base, Dir.Data, Dir.Code, Dir.Clim), function(x) if(!dir.exists(x)) dir.create(x))
```

```{r}
# read tree dataset
targets <- load("~/Desktop/Prague/mort-likelihood/individ-tree-mort/Data/full_trees.Rdata")
targets<-get(targets)
head(targets)
```

# Preparation of dataset
```{r}
# rename census date columns
colnames(targets)[names(targets) %in% "date"] <- "date_census_1"
colnames(targets)[names(targets) %in% "date_re"] <- "date_census_2"


##### combine census dates
survey_dates <- subset(targets,select=c(plotid,date_census_1,date_census_2))
survey_dates <- unique(survey_dates)

# remove NA
survey_dates <- survey_dates[!is.na(survey_dates$date_census_2),]
survey_dates <- survey_dates[order(survey_dates$plotid),]
which(duplicated(survey_dates$plotid))
summary(survey_dates)
write.csv(survey_dates,file=file.path(Dir.Data,"survey_dates.csv"),row.names=F)

# census interval summary
summary(survey_dates$date_census_2 - survey_dates$date_census_1)  # max = 7 years
```

# Load annual-resolution climate data (plot scale)
```{r}
clim_df <- load(file=file.path(Dir.Clim,"5_plot_monthly_water_budget_variables.Rdata"))
clim_df <- get(clim_df)

clim_array <- array(data=list(NULL),dim=nrow(survey_dates))

# set length of antecedent climate interval
period_length <- 30

# loop through plot and calculate quantiles
pb <- txtProgressBar(min=0, max=length(survey_dates$plotid), style=3)  # progress bar
tind=0

# loop through plots and compute mean climate for census and prior intervals
i <- 1
for(i in 1:length(survey_dates$plotid) )   
{
  tind=tind+1
  setTxtProgressBar(pb, tind)
  
  p <- survey_dates$plotid[i]
  
  # get census interval
  start_yr <- survey_dates$date_census_1[which(survey_dates$plotid %in% p)]
  end_yr <- survey_dates$date_census_2[which(survey_dates$plotid %in% p)]
  
  	# subset climate data to census interval
  	dat.plot.int <- subset(clim_df, year>=start_yr-2 &
  								year<= end_yr & # add 2 years to it
  								plotid %in% p )
  
  ### calculate mean climate for census interval
  temp_k_mean_ann_int <- mean(dat.plot.int$temp_K_ann_ave,na.rm=T)   
  temp_k_mean_seas_int <- mean(dat.plot.int$temp_K_seasonal_ave,na.rm=T)  
  WD_mm_mean_ann_int <- mean(dat.plot.int$WD_mm_ann,na.rm=T) # water deficit (PET-AET)
  WD_mm_mean_seas_int <- mean(dat.plot.int$WD_mm_seasonal,na.rm=T) # water deficit (PET-AET)
  WB_mm_mean_seas_int <- mean(dat.plot.int$water_balance_seasonal_mm,na.rm=T) # effective growing season water supply
  SMI_mm_mean_seas_int <- mean(dat.plot.int$SMI_seasonal_ave,na.rm=T) # effective growing season water supply
  ppt_mm_ann_int<- mean(dat.plot.int$ppt_mm_ann_sum,na.rm=T)
  ppt_mm_seas_int<- mean(dat.plot.int$ppt_mm_seasonal_sum,na.rm=T)
  terra_WD_mm_mean_ann_int <- mean(dat.plot.int$terra_WD_mm_ann,na.rm=T) # water deficit (PET-AET)
  terra_WD_mm_mean_seas_int <- mean(dat.plot.int$terra_WD_mm_seasonal,na.rm=T) # water deficit (PET-AET)
  terra_WB_mm_ann <- mean(dat.plot.int$WB_mm_ann,na.rm=T)
   terra_WB_mm_seas <- mean(dat.plot.int$WB_mm_seas,na.rm=T)
   
  ### extreme census climate values  (See Canham & Murphy 2017)
  temp_k_max_seas_month_int <- max(dat.plot.int$temp_K_seasonal_month_max,na.rm=T)   # max monthly_level temp during growing season
  ppt_mm_min_seas_month_int <- min(dat.plot.int$ppt_mm_seasonal_month_min,na.rm=T)  
  SMI_min_seas_month_int <- min(dat.plot.int$SMI_seasonal_month_min,na.rm=T)  
  WD_mm_max_seas_int <- max(dat.plot.int$WD_mm_seasonal,na.rm=T)
  WB_mm_min_seas_int <- min(dat.plot.int$water_balance_seasonal_mm,na.rm=T)
  
  
  ### calculate mean climate for prior interval (antecedent)
  # start_yr <- end_yr
  # end_yr <- end_yr - period_length
  
  # dat.plot.ant <- subset(clim_df, year<start_yr &
  #                          year>=end_yr &
  #                          plotid %in% p )
  dat.plot.ant <- subset(clim_df, year < 2000 &
                           					year>= 1970 &
                            					plotid %in% p)
                           
  
  temp_k_mean_seas_ant <- mean(dat.plot.ant$temp_K_seasonal_ave,na.rm=T)
  temp_k_mean_ann_ant <- mean(dat.plot.ant$temp_K_ann_ave,na.rm=T)
  WD_mm_mean_seas_ant <- mean(dat.plot.ant$WD_mm_seasonal,na.rm=T)
  WD_mm_mean_ann_ant <- mean(dat.plot.ant$WD_mm_ann,na.rm=T)
  WB_mm_mean_seas_ant <- mean(dat.plot.ant$water_balance_seasonal_mm,na.rm=T)
  ppt_mm_ann_ant<- mean(dat.plot.ant$ppt_mm_ann_sum,na.rm=T)
  ppt_mm_seas_ant<- mean(dat.plot.ant$ppt_mm_seasonal_sum,na.rm=T)
  terra_WD_mm_mean_ann_ant <- mean(dat.plot.ant$terra_WD_mm_ann,na.rm=T) # water deficit (PET-AET)
  terra_WD_mm_mean_seas_ant <- mean(dat.plot.ant$terra_WD_mm_seasonal,na.rm=T) # water deficit (PET-AET)
  
  
  ### calculate climate anomalies
  # temp_k_summer_diff <- temp_k_mean_summer_int - temp_k_mean_summer_ant 
  # WD_mm_mean_summer_diff <- WD_mm_mean_summer_int - WD_mm_mean_summer_ann_ant 
  # WB_mm_mean_summer_diff <- WB_mm_mean_seas_int - WB_mm_mean_seas_ant 
  
  # compile
  clim_array[[i]] <- data.frame("plotid"=p,  temp_k_mean_ann_int,temp_k_mean_seas_int,WD_mm_mean_seas_int,WD_mm_mean_ann_int,WB_mm_mean_seas_int,SMI_mm_mean_seas_int,ppt_mm_ann_int,ppt_mm_seas_int,temp_k_max_seas_month_int,ppt_mm_min_seas_month_int,SMI_min_seas_month_int,WD_mm_max_seas_int,WB_mm_min_seas_int, temp_k_mean_seas_ant,temp_k_mean_ann_ant,WD_mm_mean_seas_ant,WD_mm_mean_ann_ant,WB_mm_mean_seas_ant,ppt_mm_ann_ant, ppt_mm_seas_ant,terra_WD_mm_mean_ann_ant,terra_WD_mm_mean_seas_ant, terra_WD_mm_mean_ann_int, terra_WD_mm_mean_seas_int,terra_WB_mm_ann,terra_WB_mm_seas )
  
  rm(start_yr,end_yr)
}
```

# Save
```{r}
# convert to dataframe
plot_clim <- do.call(rbind,clim_array)

#summary(plot_clim)

plot_clim <- plot_clim %>%
  mutate(WD_ann_diff = WD_mm_mean_ann_int - WD_mm_mean_ann_ant,
    ppt_ann_diff = ppt_mm_ann_int - ppt_mm_ann_ant,
    temp_mean_ann_diff = temp_k_mean_ann_int - temp_k_mean_ann_ant,
    WD_ann_perc = ((WD_ann_diff/WD_mm_mean_ann_ant))*100,
    ppt_ann_perc = ((ppt_ann_diff/ppt_mm_ann_ant))*100,
    temp_mean_ann_perc = (temp_mean_ann_diff/temp_k_mean_ann_ant)*100)

# save census interval climate
save(plot_clim,file=file.path(Dir.Clim,"6_plot_mean_census_climate.Rdata"))
```
