==> Code/04_Add_Climate_to_Trees.qmd <==
---
title: "04. Tree dataset and climate combination"
format:
  html:
    theme: spacelab
    toc: true
    number-sections: true
---

# Setup
```{r}
rm(list=ls())

# user computer
#  set user computer
computer <- "~"

#  directores
if(computer == "~"){
  Dir.Base <- "~/Desktop/Prague/mort-likelihood/individ-tree-mort"
}else{
  Dir.Base <- "insert path"
}
Dir.Data <- file.path(Dir.Base, "Data")
Dir.Code <- file.path(Dir.Base, "Code")
Dir.Clim <- file.path(Dir.Data,"Climate")
Dir.Subsample <- file.path(Dir.Data, "Subsampled_data_files")
dirs <- sapply(c(Dir.Base, Dir.Data, Dir.Code,  Dir.Clim, Dir.Subsample), function(x) if(!dir.exists(x)) dir.create(x) )


library(tidyverse)
```

# Add trees
```{r}
### get tree plot data (filtered and expanded by Audrey)

targets <- load("~/Desktop/Prague/mort-likelihood/individ-tree-mort/Data/full_trees.Rdata")
targets<-get(targets)

targets$SPCD <- as.factor(targets$SPCD)
targets <- targets %>%
  rename(
    tree_status_census_1 = status,
    tree_status_census_2 = status_re,
    date_census_1 = date,
    date_census_2 = date_re
  )
```

```{r}
# length of census interval
targets$census_yrs <- targets$date_census_2 - targets$date_census_1

# dbh in cm
# targets$dbh_cm <- targets$dbh_mm / 10
# targets <- targets[,!names(targets) %in% "dbh_mm"]

#### add model response variable (mort)
# 0 = dead
# 1 = alive
targets$mort <- ifelse(targets$tree_status_census_2 %in% c(1,2,3,4),1,0)
```

```{r}

#### reclassify canopy layer data (canopy position or strata)
table(targets$layer,targets$growth)

    
  #       0    1
  # 11  274 5959
  # 12 2045 1416
  # 13 4225  890
  # 99   13    0

filter(targets,targets$layer == 99)
#filter(targets,targets$layer == -1)

# replace -1 or 99 layer values (1st census)
#targets$layer_corr <- ifelse(targets$layer %in% c(-1,99), targets$layer_re, targets$layer)

targets$layer_corr <- ifelse(targets$layer == 99 & targets$dbh_mm < 220, 13,  # assume small trees are shaded
                             ifelse(targets$layer == 99 & targets$dbh_mm >= 220, 11,
                                    targets$layer))
targets$layer_corr <- ifelse(targets$layer == -1 & targets$growth == 99, 11, targets$layer_corr)

table(targets$layer_corr,targets$growth)

# reclassify trees into two strata (upper vs lower canopy)
targets$stratum <- ifelse(targets$layer_corr %in% c(11,12),"upper","lower")
targets$stratum <- as.factor(targets$stratum)
```

# Save
```{r}
targets <- targets %>%
  filter(!stand %in% "Risnjak")
save(targets, file=file.path(Dir.Data,"8_Trees_with_mortality.Rdata"))
```

# Add Climate to tree dataset 
```{r}
# load climate
clim_df <- load(file=file.path(Dir.Clim,"6_plot_mean_census_climate.Rdata"))
clim_df <- get(clim_df)
names(clim_df)
```

```{r}
load("~/Desktop/Prague/mort-likelihood/individ-tree-mort/Data/8_Trees_with_mortality.Rdata")
# targets<-get(targets)
```

```{r}
# combine trees and climate data for period of available climate (1958-2023)
trees <- merge(targets,clim_df, by=c("plotid"), all=FALSE)
str(trees)
summary(trees)
```

### save
```{r}
save(trees,file=file.path(Dir.Data,"8_Trees_with_derived_climate_variables.Rdata"))
```
